Server/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-text-splitters

faiss-cpu
pypdf

sentence-transformers
torch

langchain-google-genai
